requirements.txt
boto3~=1.40.1
python-dotenv~=1.1.1
botocore~=1.40.1
fastapi~=0.116.1